Extract DateList and EventStatus as standalone components
