+++ content/webapp/pages/events/[eventId].tsx
@@ -1,16 +1,14 @@
 import NextLink from 'next/link';
-import { FunctionComponent, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import * as prismic from '@prismicio/client';
 import PageLayout from '@weco/common/views/components/PageLayout/PageLayout';
 import EventSchedule from '@weco/content/components/EventSchedule/EventSchedule';
-import TextWithDot from '@weco/common/views/components/TextWithDot';
 import ButtonSolid from '@weco/common/views/components/ButtonSolid/ButtonSolid';
 import ButtonSolidLink from '@weco/common/views/components/ButtonSolidLink/ButtonSolidLink';
 import EventbriteButtons from '@weco/content/components/EventbriteButtons/EventbriteButtons';
 import Message from '@weco/common/views/components/Message/Message';
 import PrismicHtmlBlock from '@weco/common/views/components/PrismicHtmlBlock/PrismicHtmlBlock';
 import InfoBox from '@weco/content/components/InfoBox/InfoBox';
-import DateRange from '@weco/common/views/components/DateRange/DateRange';
 import { font } from '@weco/common/utils/classnames';
 import { camelize } from '@weco/common/utils/grammar';
 import { formatDayDate, formatTime } from '@weco/common/utils/format-date';
@@ -57,25 +55,15 @@ import {
   prismicPageIds,
 } from '@weco/common/data/hardcoded-ids';
 import { headerBackgroundLs } from '@weco/common/utils/backgrounds';
-import { isDayPast, isPast } from '@weco/common/utils/dates';
+import { isPast } from '@weco/common/utils/dates';
+import DateList from '@weco/content/components/DateList';
+import EventStatus from '@weco/content/components/EventStatus';
 
 import * as prismicT from '@prismicio/types';
 import { JsonLdObj } from '@weco/common/views/components/JsonLd/JsonLd';
-import { PaletteColor } from '@weco/common/views/themes/config';
 import { a11y } from '@weco/common/data/microcopy';
 import { Pageview } from '@weco/common/services/conversion/track';
 import { createPrismicLink } from '@weco/common/views/components/ApiToolbar';
-
-const TimeWrapper = styled(Space).attrs({
-  v: {
-    size: 'm',
-    properties: ['padding-top', 'padding-bottom'],
-  },
-})`
-  display: flex;
-  justify-content: space-between;
-  border-top: 1px solid ${props => props.theme.color('warmNeutral.400')};
-`;
 
 const DateWrapper = styled.div.attrs({
   className: 'body-text',
@@ -98,58 +86,11 @@ const EmailTeamCopy = styled(Space).attrs({
   color: ${props => props.theme.color('neutral.700')};
 `;
 
-const DateRangeWrapper = styled.div<{ isPast: boolean }>`
-  ${props => props.isPast && `color: ${props.theme.color('neutral.600')};`};
-  flex: 1;
-`;
-
 type Props = {
   event: Event;
   jsonLd: JsonLdObj[];
   gaDimensions: GaDimensions;
   pageview: Pageview;
-};
-
-// TODO: Probably use the StatusIndicator?
-type EventStatusProps = {
-  text: string;
-  color: PaletteColor;
-};
-
-const EventStatus: FunctionComponent<EventStatusProps> = ({ text, color }) => {
-  return (
-    <div style={{ display: 'flex' }}>
-      <TextWithDot className={font('intb', 5)} dotColor={color} text={text} />
-    </div>
-  );
-};
-
-const DateList: FunctionComponent<{ event: Event }> = ({ event }) => {
-  return (
-    event.times && (
-      <>
-        {event.times.map((eventTime, index) => {
-          return (
-            <TimeWrapper key={index}>
-              <DateRangeWrapper isPast={isDayPast(eventTime.range.endDateTime)}>
-                <DateRange
-                  start={eventTime.range.startDateTime}
-                  end={eventTime.range.endDateTime}
-                />
-              </DateRangeWrapper>
-
-              {isDayPast(eventTime.range.endDateTime) ? (
-                <EventStatus text="Past" color="neutral.500" />
-              ) : eventTime.isFullyBooked.inVenue &&
-                eventTime.isFullyBooked.online ? (
-                <EventStatus text="Full" color="validation.red" />
-              ) : null}
-            </TimeWrapper>
-          );
-        })}
-      </>
-    )
-  );
 };
 
 function showTicketSalesStart(dateTime: Date | undefined) {
